Configuração para o Pytest.
