+++ app/.coverage
@@ -1,0 +1,1 @@
+!coverage.py: This is a private format, don't read it directly!{"lines":{"/home/alohomora/schema.py":[1,2,3],"/home/accounts/schema.py":[1,2,5,11,16,22],"/home/accounts/mutations/__init__.py":[1],"/home/accounts/mutations/entries.py":[1,2,3,4,5,6,7,9,10,12,13,15,16,18,19,20,21,23],"/home/condos/types.py":[1,2,4,5,6,8,9,10],"/home/accounts/types.py":[1,2,3,4,6,7,8,9,10,12,13,14,15,16,18,19,20,21,22,24,25,26,27,28,30,31,32,33,34,36,37,38,39,40,42,43,44,45,46,48,49,50,51,52,54,55,56,57,58,59,60,61,62,63,64,66,67,68,69,70],"/home/accounts/mutations/entries_visitors.py":[1,2,3,4,5,6,7,9,10,12,14,15,16,17,18,19,21,57,59,60,62,63,65,79,80,82,84,85,86,88,97,98,100,102,103,104,106],"/home/accounts/mutations/residents.py":[1,2,3,4,5,6,8,9,11,13,14,15,16,17,18,19,20,21,26,27,29,30,79,80,81,82,84,85,86,89,109,110,111,113,114,115,118],"/home/accounts/mutations/services.py":[1,2,3,4,5,7,8,10,12,13,14,15,16,19,38,39,40,41,43,44,45,48,68,69,70,72,73,74,77],"/home/accounts/mutations/users.py":[1,2,3,4,6,16],"/home/core/models.py":[],"/home/core/tests.py":[],"/home/core/__init__.py":[],"/home/core/views.py":[],"/home/core/admin.py":[],"/home/core/management/__init__.py":[],"/home/core/management/commands/__init__.py":[],"/home/core/management/commands/loadinitialdata.py":[],"/home/alohomora/__init__.py":[],"/home/alohomora/runner.py":[],"/home/alohomora/wsgi.py":[],"/home/accounts/models.py":[],"/home/accounts/__init__.py":[],"/home/accounts/views.py":[],"/home/accounts/utility.py":[],"/home/accounts/admin.py":[],"/home/accounts/mutations/visitors.py":[],"/home/accounts/mutations/admins.py":[],"/home/bot/models.py":[],"/home/bot/__init__.py":[],"/home/bot/test_bot.py":[],"/home/bot/views.py":[],"/home/bot/schema.py":[],"/home/bot/admin.py":[],"/home/condos/models.py":[],"/home/condos/__init__.py":[],"/home/condos/views.py":[],"/home/condos/schema.py":[],"/home/condos/test_condos.py":[],"/home/condos/queries.py":[],"/home/condos/admin.py":[],"/home/condos/mutations/apartments.py":[],"/home/condos/mutations/__init__.py":[],"/home/condos/mutations/blocks.py":[]}}
