+++ app/alohomora/runner.py
@@ -1,0 +1,42 @@
+'''
+Customize ./manage.py test command to pytest
+'''
+
+
+class PytestTestRunner:
+    '''
+    Runs pytest to discover and run tests.
+    '''
+    # pylint: disable=too-few-public-methods
+
+    def __init__(
+            self,
+            verbosity=1,
+            failfast=False,
+            keepdb=False,
+            **unused_kwargs):
+        self.verbosity = verbosity
+        self.failfast = failfast
+        self.keepdb = keepdb
+
+    def run_tests(self, test_labels):
+        """
+        Run pytest and return the exitcode.
+        It translates some of Django's test command option to pytest's.
+        """
+        import pytest
+
+        argv = []
+        if self.verbosity == 0:
+            argv.append('--quiet')
+        if self.verbosity == 2:
+            argv.append('--verbose')
+        if self.verbosity == 3:
+            argv.append('-vv')
+        if self.failfast:
+            argv.append('--exitfirst')
+        if self.keepdb:
+            argv.append('--reuse-db')
+
+        argv.extend(test_labels)
+        return pytest.main(argv)
