+++ app/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = alohomora.settings
+#addopts = --nomigrations --cov=. --cov-report term-missing
