+++ docker/requirements.txt
@@ -23,3 +23,7 @@ urllib3==1.25.3
 django-graphql-jwt==0.2.1
 python_speech_features==0.6
 fastdtw==0.3.2
+pytest-django==3.7.0
+pytest==5.2.2
+pytest-cov==2.8.1
+snapshottest==0.5.1
